my-auth-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
psycopg2-binary
python-dotenv
